fix(feed): use active positions + centralize narrative scoring
